pyproject.toml
[tool.poetry]
name = "shotbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "NONE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"
tk = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
